Add slash to the line-numbered edit buttons
The edit URL on the edit buttons on each headline in a compendium
has no slash before the hash symbol.
As the edit URL ends with a slash, browsers redirect to a URL with
the slash.

In safari this redirection causes the hash part of the URL to disappear.

This fix avoids the redirect by adding a slash to the edit URLs.