Redirect article to correct language
Today, a user that goes to wikipendium.no/TEP4105 will be presented
with the missing language screen, as the compendium only exist in
norwegian.
This makes the use of short URLs to compendiums impractical.

With this pull request, if the requested language version doesn't exist,
the user will be redirected a version that does exist, if there is only
one language version of the article.

If there exist multiple, the user will still be presented
with the existing 'missing language' page.

Closes #220