Disable caching of redirects in article view
That turned out to be a bad idea.

If a user first goes to the URL wikipendium.no/tdt4258,
the cache will remember that the view for the article tdt4258
should always redirect to wikipendium.no/TDT4258_Energy...

Since the actual path is not part of the cache key, this causes a
redirect loop.