Fix syntax highlighting issues in codemirror
This updates codemirror to the latest version and closes #230

It will also properly respect escaped dollar signs.

It takes advantage of the more appropriate multiplexing mode in
codemirror. This means that while in math mode the markdown parser will
ignore all characters. With the overlay mode we're currently using, a
"_" or "*" character inside math will cause the markdown syntax
highlighting to activate italic text. Multiplexing resolves this issue.

Finally, it introduces highlighting of latex/math, so it's easier to
spot syntax issues here as well, and less green overall.