Redesign table of contents to a left-anchored menu
It is now hidden on mobile, with expandable headlines instead.
The font size on mobile is reduced slightly, so headlines and text in
general will more often fit on one line.

The typography of the article title is also changed, to differentiate it
from the other headlines now that they are placed more closesly
together.

Fixes #38