Fix page-footer height
The sticky footer implementation we currently use depends on a
fixed-height footer. Ever since we added the paypal donation to the
footer, this fixed height has been too small, resulting in scrollbars on
even very short articles.

This restores the footer back to being perfectly on the bottom of the
screen on pages with little content.