Add django-haystack/whoosh-based full-text search
The search is available at //BASEURL/search/. The search uses
django-haystack with a Whoosh backend. A cronjob must perform periodical
search index updates for this to work. Since we want to prune old
article contents, and not allow people to search in old revisions of
articles (for UX simplicity's sake), we might as well just rebuild the
entire index on every update instead of incrementally updating it, as is
normally done. A suggest interval for updating the search index is once
every five minutes. I made this number up right now, but it seems like a
reasonable choice. The alternative is to reactively rebuild search
indexes upon article content creation, but is more complex to set up.

The command for rebuilding the search index is
`python manage.py rebuild_index`.