Add JITishCron
JITishCron (Just-In-Time-ish Cron) is a request-driven task runner with
a not-more-often-than-x execution semantic. Tasks can be specified to be
run on each page load, but never more often than every x seconds. This
is useful for tasks that would usually be run by some external timer,
e.g. cron, but that do not really need to be run unless some activity
has happened on the site. Taking a request/page-load as a proxy for
meaningful activity, JITishCron caters to this use case.

JITishCron tasks are synchronously executed "just in time" for view
processing. To reduce page load times, JITishCron tasks should therefore
be lightweight or run its computation asynchronously, e.g. by using
subprocess.Popen. Doing things asynchronously means that the computation
will no longer be completed "in time" for view processing for the
triggering request, hence the "ish" part of "JITishCron".

JITishCron relies on unique constraints in the database layer to ensure
that multiple workers do not execute the same task at the same time,
violating the not-more-often-than-x semantic. This means that if you are
using a database that does not enforce unique constraints globally,
perhaps because it is not implemented at all, or because the database is
massively distributed, then JITishCron will probably break.

JITishCron does not verifiy the execution status of executed tasks
(i.e. results, return values, exit codes etc.).