Add automatic .scss static file loader
Now, all findable .scss static files will be automatically added to
base.html. This means that .scss files do not need to be manually linked
using HTML <link> tags.

The templatetag for this was put in the wiki app for now, even though it
does not directly pertain to it.