Correct base template path in registration view
We don't currently use this template, as we redirect users to the front
page after signup instead. But for a while the frontpage redirect hasn't
been working, which means that we've been trying to render this template
instead, because this template is the default endpoint for user signup.

Because of this faulty path, we've served a sexy 500 server error to
every new wikipendium user.