Add data migration to uppercase all article slugs
We currently assume that all article slugs are uppercase. However, prior
to this commit, it was possible to create articles with lowercase slugs
using the UI, which means that databases could at the point of this data
contain invalid data, i.e. lowercased slugs. This commit introduces a
data migration that uppercases all slugs, ensuring that the database is
returned to a valid state.