Add error logging to file
This splits the settings so that we can have different logging for base,
test and local.

Also, django-appconf is an indirect dependency of django-compressor
which apparently has released a patch version that drops python2 support
since the last time we did anything (which reminds me that we really
should not be running this on python 2 in prod much longer!) Anyway. I
had to pin it to the previous version to keep this working.